requirements.txt
requests
python-dotenv
colorama
websockets
